Se modifica schema

+++ microservicio/src/main/resources/db/migration/DDL/V1.0__schema.sql
@@ -10,6 +10,6 @@ create table producto (
  id int(11) not null auto_increment,
  nombre varchar(100) not null,
  descripcion varchar(100) null,
- precio double(11,2) not null,
+ precio decimal(11,2) not null,
  primary key (id)
 );
